api/requirements.txt
accelerate==0.16.0
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
diffusers==0.12.1
fastapi==0.92.0
filelock==3.9.0
h11==0.14.0
httptools==0.5.0
huggingface-hub==0.12.0
idna==3.4
importlib-metadata==6.0.0
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
psutil==5.9.4
pydantic==1.10.5
python-dotenv==0.21.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
scipy==1.10.0
sniffio==1.3.0
starlette==0.25.0
tokenizers==0.13.2
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.64.1
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
zipp==3.13.0
